gmx_clean_20250918_100616/backend/python_jobs/orquestador/requirements.txt
functions-framework==3.*
google-auth==2.*
google-cloud-bigquery==3.*
google-api-core==2.*
requests==2.*
